[FEATURE] Gerenciamento de perfils x usuários x permissões.

+++ odonto-delta-api/src/main/java/br/com/molens/odontoDelta/gateway/entity/PerfilUsuarioPK.java
@@ -19,9 +19,9 @@ import lombok.NoArgsConstructor;
 public class PerfilUsuarioPK implements Serializable {
 
     @Column(name = "PF_ID", nullable = false)
-    private Integer idPerfil;
+    private Long idPerfil;
 
     @Column(name = "US_ID", nullable = false)
-    private Integer idUsuario;
+    private Long idUsuario;
 
 }
